requirements.txt
fastapi
uvicorn
SQLAlchemy
pymysql
aiomysql
python-dotenv
cryptography